refactor(06-02): extract write-contract core logic to write-contract-core.ts
- Create write-contract-core.ts with writeContractCore, WriteContractCoreInput, WriteContractResult
- Move stepHandler, WriteContractResult type, and WriteContractCoreInput type to core file
- Update write-contract.ts to import from core and delegate to writeContractCore
- No "use step" directive in core file -- safe to import from any future step
